Moved functions from service_search to ansible_ops
